icaltz-util: Correct 64-bit value decode from tzfile on little endian
This can be seen for example on Asia/Kolkata, where the transition time
starts back at 1854-06-27. When this negative value is type-cast
to int it looses upper bytes and truncates the value to smaller,
into year 2006 instead.

+++ src/libical/icaltz-util.c
@@ -168,7 +168,7 @@ static long long int decode64(const void *ptr)
     if ((BYTE_ORDER == BIG_ENDIAN)) {
         return *(const long long int *)ptr;
     } else {
-        return (int)bswap_64(*(const unsigned long long int *)ptr);
+        return (long long int)bswap_64(*(const unsigned long long int *)ptr);
     }
 #endif
 }
